prism_code/requirements.txt
xarray
geopandas
shapely
rioxarray
requests
python-dateutil
affine
glob2
